# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Intelligence Dashboard API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Article {
  id: number;
  title: string;
  source: string;
  publishedAt: string;
  url: string;
  /** @nullable */
  rawContent?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  eventType?: string | null;
  /** @nullable */
  sentiment?: string | null;
  /** @nullable */
  whyItMatters?: string | null;
  /** @nullable */
  whoCares?: string | null;
  cloImpact: boolean;
  /** @nullable */
  issuerName?: string | null;
  /** @nullable */
  urgencyScore?: number | null;
  covenantFlag: boolean;
  /** @nullable */
  ratingMentioned?: string | null;
  /** @nullable */
  ratingAgency?: string | null;
  /** @nullable */
  marketImpact?: string | null;
  /** @nullable */
  processedAt?: string | null;
  createdAt: string;
}

export interface ArticleList {
  articles: Article[];
  total: number;
}

export interface SectorSignal {
  sector: string;
  totalArticles: number;
  negativeCount: number;
  eventTypes: string[];
  riskScore: number;
}

export interface EventTypeSignal {
  eventType: string;
  count: number;
  negativeCount: number;
  sectors: string[];
}

export interface IssuerSignal {
  issuerName: string;
  /** @nullable */
  sector?: string | null;
  totalArticles: number;
  negativeCount: number;
  covenantFlag: boolean;
  maxUrgency: number;
  eventTypes: string[];
  /** @nullable */
  ratingMentioned?: string | null;
  /** @nullable */
  ratingAgency?: string | null;
  /** @nullable */
  marketImpact?: string | null;
  latestArticleDate: string;
  riskScore: number;
}

export interface IssuerList {
  issuers: IssuerSignal[];
  total: number;
}

export interface SignalsResponse {
  bySector: SectorSignal[];
  byEventType: EventTypeSignal[];
  totalArticles: number;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface DailyBriefItem {
  articleId: number;
  title: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  sentiment?: string | null;
  /** @nullable */
  eventType?: string | null;
  /** @nullable */
  issuerName?: string | null;
  /** @nullable */
  urgencyScore?: number | null;
  covenantFlag: boolean;
  /** @nullable */
  ratingMentioned?: string | null;
  /** @nullable */
  ratingAgency?: string | null;
  /** @nullable */
  marketImpact?: string | null;
}

export interface DailyBrief {
  date: string;
  mostNegativeEvents: DailyBriefItem[];
  mostImpactedSectors: SectorSignal[];
  keyTrends: string[];
  cloAlerts: DailyBriefItem[];
  covenantAlerts: DailyBriefItem[];
  criticalAlerts: DailyBriefItem[];
  totalArticlesProcessed: number;
}

export interface RefreshResult {
  fetched: number;
  processed: number;
  duplicatesSkipped: number;
  errors: number;
  message: string;
}

export type ListArticlesParams = {
  sector?: string;
  eventType?: string;
  sentiment?: string;
  issuerName?: string;
  covenantFlag?: boolean;
  marketImpact?: string;
  minUrgency?: number;
  limit?: number;
  offset?: number;
};