# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Credit Intelligence Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List all processed articles
      parameters:
        - name: sector
          in: query
          required: false
          schema:
            type: string
        - name: eventType
          in: query
          required: false
          schema:
            type: string
        - name: sentiment
          in: query
          required: false
          schema:
            type: string
        - name: issuerName
          in: query
          required: false
          schema:
            type: string
        - name: covenantFlag
          in: query
          required: false
          schema:
            type: boolean
        - name: marketImpact
          in: query
          required: false
          schema:
            type: string
        - name: minUrgency
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Article not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        source:
          type: string
        publishedAt:
          type: string
          format: date-time
        url:
          type: string
        rawContent:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
        whyItMatters:
          type: ["string", "null"]
        whoCares:
          type: ["string", "null"]
        cloImpact:
          type: boolean
        issuerName:
          type: ["string", "null"]
        urgencyScore:
          type: ["integer", "null"]
        covenantFlag:
          type: boolean
        ratingMentioned:
          type: ["string", "null"]
        ratingAgency:
          type: ["string", "null"]
        marketImpact:
          type: ["string", "null"]
        processedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - source
        - publishedAt
        - url
        - cloImpact
        - covenantFlag
        - createdAt
    ArticleList:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
      required:
        - articles
        - total
...