# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Intelligence Dashboard API
 * OpenAPI spec version: 0.2.0
 */

export interface Article {
  id: number;
  title: string;
  source: string;
  publishedAt: Date;
  url: string;
  /** @nullable */
  rawContent?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  eventType?: string | null;
  /** @nullable */
  sentiment?: string | null;
  /** @nullable */
  whyItMatters?: string | null;
  /** @nullable */
  whoCares?: string | null;
  cloImpact: boolean;
  /** @nullable */
  issuerName?: string | null;
  /** @nullable */
  urgencyScore?: number | null;
  covenantFlag: boolean;
  /** @nullable */
  ratingMentioned?: string | null;
  /** @nullable */
  ratingAgency?: string | null;
  /** @nullable */
  marketImpact?: string | null;
  /** @nullable */
  processedAt?: Date | null;
  createdAt: Date;
}