# lib/api-zod/src/generated/types/articleList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Intelligence Dashboard API
 * OpenAPI spec version: 0.2.0
 */
import type { Article } from "./article";

export interface ArticleList {
  articles: Article[];
  total: number;
}